requirements.txt
Distance==0.1.3
jsonlines==4.0.0
lmdb==1.5.1
matplotlib==3.7.5
mmcv==2.2.0
mmdet==3.3.0
numpy==1.22.4
numpy==1.24.4
opencv_python==4.8.1.78
pandas==2.0.3
Pillow==11.1.0
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
timm==1.0.15
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.2
zhconv==1.4.3